+++ examples/option_groups.py
@@ -63,4 +63,4 @@ def main(**kwargs):
 
 
 if __name__ == '__main__':
-    main()
+    main('--help'.split())